Discography/requirements.txt
requests>=2.31.0
cryptography>=42.0.0
pystray>=0.19.5
Pillow>=10.3.0
pyinstaller>=6.6.0
